Added checking and applying of clang-format
Add a convenience script to allow checking of clang format on files in
the repo. This adds a bit nicer error reporting on top of
'git clang-format', and works in CI where 'git clang-format' may not be
available.

We also add the github workflow to make sure that clang format is
applied.

Usage of the script in CI is:

  tools/run-clang-format.py check

and to apply clang-format to all .c, .cpp, .h and .hpp files, use

  tools/run-clang-format.py apply