Update script to run clang-format
We need to be able to run the same clang-format as in CI, as otherwise
we may get different results. As such, the user can now pass the
executable for clang-format directly to the script, so that a different
version from the one on the PATH can be found.

The CI currently uses clang 14.0, which is installed using
`apt install clang-format`